Refactor: moved rotation logic to a service.ts file
